Planner extensions - delete cstor
We now generate a cstor on all complex types to initialize odata.type. We have a cstor clash with planner extensions so we need to delete the planner cstors that initiialized the AdditionalData property bag. Didn't want to initialize AddtionalData in the generated cstor. We are now null checking and creating AdditionalData before access.